Cleaned up changes to GridAccessor and formatting
