+++ test/celeritas/optical/ValidationUtils.hh
@@ -8,7 +8,6 @@
 #pragma once
 
 #include <array>
-#include <type_traits>
 #include <vector>
 
 #include "corecel/cont/Array.hh"
@@ -22,6 +21,8 @@
 #include "Test.hh"
 #include "TestMacros.hh"
 
+#include "detail/ValidationUtilsImpl.hh"
+
 //! Equivalence macro for physics grids
 #define EXPECT_GRID_EQ(expected, actual) \
     EXPECT_PRED_FORMAT2(::celeritas::testdetail::IsGridEq, expected, actual)
@@ -32,133 +33,6 @@
 
 namespace celeritas
 {
-namespace testdetail
-{
-//---------------------------------------------------------------------------//
-/*!
- * Type traits for physics grids.
- *
- * Allows duck-typing to allow comparisons of physics grids that might be
- * stored as different objects.
- */
-template<class GridType>
-struct PhysicsGridTraits;
-
-//---------------------------------------------------------------------------//
-//! Specialization for \c ImportPhysicsVector
-template<>
-struct PhysicsGridTraits<ImportPhysicsVector>
-{
-    using grid_type = ImportPhysicsVector;
-
-    static constexpr std::vector<double> const& grid(grid_type const& v)
-    {
-        return v.x;
-    }
-    static constexpr std::vector<double> const& value(grid_type const& v)
-    {
-        return v.y;
-    }
-};
-
-//---------------------------------------------------------------------------//
-//! Specialization for a tuple of containers
-template<class T>
-struct PhysicsGridTraits<std::tuple<T, T>>
-{
-    using grid_type = std::tuple<T, T>;
-    static constexpr T const& grid(grid_type const& v)
-    {
-        return std::get<0>(v);
-    }
-    static constexpr T const& value(grid_type const& v)
-    {
-        return std::get<1>(v);
-    }
-};
-
-//---------------------------------------------------------------------------//
-/*!
- * Compare to physics grids with exact equivalence.
- */
-template<class GridTypeE, class GridTypeA>
-::testing::AssertionResult IsGridEq(char const* expected_expr,
-                                    char const* actual_expr,
-                                    GridTypeE const& expected,
-                                    GridTypeA const& actual)
-{
-    using EGT = PhysicsGridTraits<GridTypeE>;
-    using AGT = PhysicsGridTraits<GridTypeA>;
-
-    auto x_result = IsVecEq(
-        expected_expr, actual_expr, EGT::grid(expected), AGT::grid(actual));
-    auto y_result = IsVecEq(
-        expected_expr, actual_expr, EGT::value(expected), AGT::value(actual));
-
-    ::testing::AssertionResult result(x_result && y_result);
-    if (!x_result)
-    {
-        result << "x values:\n" << x_result.message();
-    }
-    if (!y_result)
-    {
-        result << "y values:\n" << y_result.message();
-    }
-
-    return result;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Compare physics tables with exact equivalence.
- */
-template<class GridTypeE, class GridTypeA>
-::testing::AssertionResult IsTableEq(char const* expected_expr,
-                                     char const* actual_expr,
-                                     std::vector<GridTypeE> const& expected,
-                                     std::vector<GridTypeA> const& actual)
-{
-    if (expected.size() != actual.size())
-    {
-        ::testing::AssertionResult failure = ::testing::AssertionFailure();
-
-        failure << " Size of: " << actual_expr
-                << "\n  Actual: " << actual.size()
-                << "\nExpected: " << expected_expr
-                << ".size()\nWhich is: " << expected.size() << "\n";
-        return failure;
-    }
-
-    ::testing::AssertionResult result = ::testing::AssertionSuccess();
-
-    for (auto i : range(expected.size()))
-    {
-        std::string index_expr = "[" + std::to_string(i) + "]";
-        std::string expected_expr_i = expected_expr + index_expr;
-        std::string actual_expr_i = actual_expr + index_expr;
-
-        auto grid_result = IsGridEq(expected_expr_i.c_str(),
-                                    actual_expr_i.c_str(),
-                                    expected[i],
-                                    actual[i]);
-
-        if (!grid_result)
-        {
-            if (result)
-            {
-                result = ::testing::AssertionFailure();
-            }
-
-            result << grid_result.message();
-        }
-    }
-
-    return result;
-}
-
-//---------------------------------------------------------------------------//
-}  // namespace testdetail
-
 namespace optical
 {
 namespace test
@@ -179,6 +53,20 @@ Array<real_type, sizeof...(Args)> constexpr native_array_from(Args const&... arg
 
 //---------------------------------------------------------------------------//
 /*!
+ * Helper function to annotate units of a hard-coded test data array.
+ *
+ * This overload accepts fixed size \c Arrays which clang-format handles
+ * better.
+ */
+template<class UnitType, size_t N>
+Array<real_type, N> constexpr native_array_from(Array<real_type, N> const& arr)
+{
+    return ::celeritas::testdetail::native_array_from<UnitType>(
+        arr, std::make_index_sequence<N>{});
+}
+
+//---------------------------------------------------------------------------//
+/*!
  * Helper function to annotate units of hard-coded test data.
  *
  * Same as \c native_array_from, except returns a vector.
@@ -191,9 +79,10 @@ std::vector<real_type> native_vector_from(Args const&... args)
 
 //---------------------------------------------------------------------------//
 /*!
- * Perform consistency checks on grids built in \c Collections.
+ * Convenience class for accessing data built on the grid, and performing
+ * sanity checks on bounds.
  */
-class GridValidator
+class GridAccessor
 {
   public:
     //!@{
@@ -201,39 +90,22 @@ class GridValidator
     using Grid = GenericGridRecord;
     using GridId = OpaqueId<Grid>;
     using ImportPhysicsTable = std::vector<ImportPhysicsVector>;
+    using GridView = std::tuple<Span<real_type const>, Span<real_type const>>;
 
     template<class T>
     using Items = Collection<T, Ownership::value, MemSpace::host>;
     //!@}
 
   public:
-    // Construct validator for underlying storage
-    GridValidator(Items<real_type>* reals, Items<Grid>* grids);
+    // Construct accessor for underlying storage
+    GridAccessor(Items<real_type>* reals, Items<Grid>* grids);
 
-    std::vector<std::tuple<Span<real_type const>, Span<real_type const>>>
-    operator()(ItemRange<Grid> grid_ids) const
-    {
-        std::vector<std::tuple<Span<real_type const>, Span<real_type const>>> grids;
-        grids.reserve(grid_ids.size());
+    // Retrieve a table of grid views built on the storage
+    std::vector<GridView> operator()(ItemRange<Grid> grid_ids) const;
 
-        for (GridId grid_id : grid_ids)
-        {
-            CELER_EXPECT(grid_id < grids_->size());
-            auto const& grid = (*grids_)[grid_id];
-            CELER_EXPECT(grid);
-            grids.push_back(
-                std::make_tuple((*this)(grid.grid), (*this)(grid.value)));
-        }
-
-        return grids;
-    }
-
-    Span<real_type const> operator()(ItemRange<real_type> const& real_ids) const
-    {
-        CELER_EXPECT(real_ids.front() < real_ids.back());
-        CELER_EXPECT(real_ids.back() < reals_->size());
-        return (*reals_)[real_ids];
-    }
+    // Retrieve a span of reals built on the storage
+    Span<real_type const>
+    operator()(ItemRange<real_type> const& real_ids) const;
 
     // Construct an MFP builder with the underlying collections
     MfpBuilder create_mfp_builder();
@@ -245,13 +117,13 @@ class GridValidator
 
 //---------------------------------------------------------------------------//
 /*!
- * A \c GridValidator that stores its own collections.
+ * A \c GridAccessor that stores its own collections.
  */
-class OwningGridValidator : public GridValidator
+class OwningGridAccessor : public GridAccessor
 {
   public:
     // Construct with internal collections
-    OwningGridValidator();
+    OwningGridAccessor();
 
   private:
     Items<real_type> reals_;
